Moves (all) logic from turnofflayers.js to legend.js and exempts group:none layers
