+++ src/controls.js
@@ -20,4 +20,3 @@ export { default as Sharemap } from './controls/sharemap';
 export { default as Splash } from './controls/splash';
 export { default as Zoom } from './controls/zoom';
 export { default as Externalurl } from './controls/externalurl';
-export { default as Turnofflayers } from './controls/turnofflayers';
